+++ index.php
@@ -6,17 +6,7 @@ error_reporting (1);
 include("config.php");
 include("functions.php");
 session_start();
-require_once(__ROOT__.'/html/models/relay.php'); 
-
-/**
-index.php File
-
-@ Realityhost.nl
-
-Created at 7-1-2009  (dd/mm/yyyy) at 11:00
-**/
-
-
+require_once(__ROOT__.'/html/models/model.php'); 
 
 /** Als er geen sessie is ga je naar de login pagina **/
 if(empty($_SESSION['username']) OR empty($_SESSION['password'])) {
